# ⚡️ Faster implementation for `safeApply`