--- a/.yarn/versions/2eb7f70f.yml
+++ b/.yarn/versions/2eb7f70f.yml
@@ -1,0 +1,6 @@
+releases:
+  fast-check: patch
+
+declined:
+  - "@fast-check/ava"
+  - "@fast-check/jest"