--- a/packages/fast-check/src/utils/apply.ts
+++ b/packages/fast-check/src/utils/apply.ts
@@ -1,9 +1,20 @@
-const safeObjectGetPrototypeOf = Object.getPrototypeOf;
-const safeObjectGetOwnPropertyDescriptor = Object.getOwnPropertyDescriptor;
-const untouchedPrototype = Function.prototype;
 const untouchedApply = Function.prototype.apply;
-
 const ApplySymbol = Symbol('apply');
+
+/**
+ * Extract apply or return undefined
+ * @param f - Function to extract apply from
+ * @internal
+ */
+function safeExtractApply<T, TArgs extends unknown[], TReturn>(
+  f: (this: T, ...args: TArgs) => TReturn
+): ((thisArg: T) => TReturn) | undefined {
+  try {
+    return f.apply;
+  } catch (err) {
+    return undefined;
+  }
+}
 
 /**
  * Equivalent to `f.apply(instance, args)` but temporary altering the instance
@@ -30,12 +41,10 @@ export function safeApply<T, TArgs extends unknown[], TReturn>(
   instance: T,
   args: TArgs
 ): TReturn {
-  const fPrototype = safeObjectGetPrototypeOf(f);
-  if (fPrototype === untouchedPrototype && safeObjectGetOwnPropertyDescriptor(f, 'apply') === undefined) {
-    const functionApplyDesc = safeObjectGetOwnPropertyDescriptor(untouchedPrototype, 'apply');
-    if (functionApplyDesc !== undefined && functionApplyDesc.value === untouchedApply) {
-      return f.apply(instance, args);
-    }
+  // Not as safe as checking the descriptor of the property but much faster
+  // Can be by-passed by an appropriate getter property on 'apply'
+  if (safeExtractApply(f) === untouchedApply) {
+    return f.apply(instance, args);
   }
   return safeApplyHacky(f, instance, args);
 }