@dubzzz 2022-08-31
Performance measurements for current main https://github.com/dubzzz/fast-check/commit/703e44e5e4532a9e91b72d63dd9f00fb90800364 (<em>aka main</em>) vs 3.1.0 (<em>aka 3.1.0</em>) vs this PR (<em>aka extra</em>):</p>


...

More details at: https://github.com/dubzzz/fast-check-benchmarks/actions/runs/2960555414